+++ src/api/league-game/index.ts
@@ -1,4 +1,4 @@
-import { useGetAllGameList } from './use-get-all-game-list';
+import { useGetAllGameList } from '../league/use-get-all-game-list';
 import { useGetGameListByDate } from './use-get-game-list-by-date';
 import { useGetGameByLeagueGameId } from './use-get-game-by-league-game-id';
 import { useGetGameListByLeagueGameIdYearMonth } from './use-get-game-list-by-league-game-id-year-month';
