+++ src/api/league/index.ts
@@ -1,0 +1,4 @@
+import { useGetAllGameList } from '@/api/league/use-get-all-game-list';
+import { useGetTeamList } from '@/api/league/use-get-team-list';
+
+export { useGetTeamList, useGetAllGameList };
