+++ src/components/league-game/team-logo.tsx
@@ -1,10 +1,9 @@
 import Image from 'next/image';
 import { twMerge } from 'tailwind-merge';
-import cimg from '@/assets/Chelsea.png';
 
 type TeamLogoProps = {
   name: string;
-  // logo: string;
+  logo: string;
   teamId: number;
   onSelectedTeamId: (teamId: number) => void;
   selectedTeamId: number;
@@ -14,25 +13,29 @@ export default function TeamLogo({
   selectedTeamId,
   onSelectedTeamId,
   name,
-  // logo,
+  logo,
   teamId,
 }: TeamLogoProps) {
-  // Todo: 로고 받아야함
-
   return (
     <div className="flex h-20 w-24 flex-col items-center">
       <button
         className={twMerge(
-          'relative my-auto w-full overflow-hidden rounded-md transition-colors hover:bg-white/60',
+          'relative my-auto h-full w-full overflow-hidden rounded-md transition-colors hover:bg-white/60',
           selectedTeamId === teamId ? 'bg-white text-blue-700' : '',
         )}
         onClick={() => onSelectedTeamId(teamId)}
         type="button"
       >
         <div className="object-contain">
-          <Image alt={name} className="mx-auto" height={60} src={cimg} width={60} />
+          <Image
+            alt={name}
+            className="mx-auto h-[50px] w-[50px]"
+            height={40}
+            src={logo}
+            width={40}
+          />
         </div>
-        <span className="text-center text-xs font-semibold">{name}</span>
+        <p className="truncate px-2 text-center text-xs font-semibold">{name}</p>
       </button>
     </div>
   );
