+++ src/types/league-games.ts
@@ -1,3 +1,5 @@
+import { TAllGameList } from '@/types/leagues';
+
 export type TGame = {
   leagueGameId: number;
   date: string;
@@ -52,74 +54,6 @@ export type TGameListByLeagueIdYearMonth = {
 
 export type TGameListByDate = {
   responses: TGame[];
-};
-
-export type TAllGameList = {
-  findLeagueGames: {
-    leagueGameId: number;
-    gameDate: string;
-    count: number;
-    away: string;
-    home: string;
-    score: {
-      homeScore: string;
-      awayScore: string;
-    };
-    leagueGameStatus: string;
-    homePlayers:
-      | {
-          LeagueGamePlayerId: number;
-          status: string;
-          playedTime: number;
-          subTime: number;
-          position: string;
-          player: {
-            playerId: number;
-            national: string;
-            playerName: string;
-            position: string;
-            leagueTeam: {
-              leagueTeamId: number;
-              leagueTeamName: string;
-              teamType: string;
-              league: string;
-              season: string;
-              logo: string;
-            };
-            birth: string;
-            season: string;
-          };
-        }[]
-      | null;
-    awayPlayers:
-      | {
-          LeagueGamePlayerId: number;
-          status: string;
-          playedTime: number;
-          subTime: number;
-          position: string;
-          player: {
-            playerId: number;
-            national: string;
-            playerName: string;
-            position: string;
-            leagueTeam: {
-              leagueTeamId: number;
-              leagueTeamName: string;
-              teamType: string;
-              league: string;
-              season: string;
-              logo: string;
-            };
-            birth: string;
-            season: string;
-          };
-        }[]
-      | null;
-  }[];
-
-  totalElements: number;
-  totalPages: number;
 };
 
 export type TGameListByLeagueTeamId = TAllGameList;
